refactor player view model and http service
