+++ XamTwitch/XamTwitch/ViewModels/PlayerPageViewModel.cs
@@ -3,15 +3,19 @@ using System.Net.Http;
 using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.MobCAT;
 using Microsoft.MobCAT.MVVM;
 using Newtonsoft.Json;
 using Xamarin.Essentials;
 using XamTwitch.Models;
+using XamTwitch.Services;
 
 namespace XamTwitch.ViewModels
 {
     public class PlayerPageViewModel : BaseNavigationViewModel
     {
+        private readonly ITwitchHttpService _twitchHttpService;
+
         private TwitchStream _stream;
         public TwitchStream Stream
         {
@@ -34,6 +38,7 @@ namespace XamTwitch.ViewModels
 
         public PlayerPageViewModel()
         {
+            _twitchHttpService = ServiceContainer.Resolve<ITwitchHttpService>();
         }
 
         private async Task FetchTwitchStreamAsync()
@@ -43,37 +48,7 @@ namespace XamTwitch.ViewModels
                 if (Stream == null)
                     return;
 
-                System.Diagnostics.Debug.WriteLine($"{Thread.CurrentThread.ManagedThreadId}, IsMain: {MainThread.IsMainThread}");
-                var channelName = Stream.UserName.ToLower();
-                var httpClient = new HttpClient();
-                httpClient.DefaultRequestHeaders.TryAddWithoutValidation("Client-ID", Constants.TwitchApiKey);
-                var tokenUrl = $"https://api.twitch.tv/api/channels/{channelName}/access_token";
-                var tokenResponse = await httpClient.GetStringAsync(tokenUrl).ConfigureAwait(false);
-                var token = JsonConvert.DeserializeObject<TwitchToken>(tokenResponse);
-
-                if (string.IsNullOrWhiteSpace(token.Sig) || string.IsNullOrWhiteSpace(token.Token))
-                    throw new InvalidOperationException();
-
-                var tokenUpdated = Uri.EscapeUriString(token.Token);
-                System.Diagnostics.Debug.WriteLine($"Original: {token.Token}\nUpdated: {tokenUpdated}");
-
-                var playlistUrl = $"https://usher.ttvnw.net/api/channel/hls/{channelName}.m3u8?sig={token.Sig}&token={tokenUpdated}";
-                System.Diagnostics.Debug.WriteLine(playlistUrl);
-
-                var playlistReponse = await httpClient.GetStringAsync(playlistUrl).ConfigureAwait(false);
-
-                var match = Regex.Match(playlistReponse, @"(?<stream>https://(.+).m3u8)");
-                if(!match.Success)
-                    throw new InvalidOperationException();
-
-                var streamUrl = match.Groups["stream"].Value;
-
-                System.Diagnostics.Debug.WriteLine($"{Thread.CurrentThread.ManagedThreadId}, IsMain: {MainThread.IsMainThread}");
-                // need UI
-                MainThread.BeginInvokeOnMainThread(() =>
-                {
-                    StreamSource = streamUrl;
-                });
+                StreamSource = await _twitchHttpService.GetTwitchStreamUrlAsync(Stream.UserName);
             }
             catch (Exception ex)
             {
